+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changed
 ### Removed
 - remove Python 3.8 support
+### Added
+- All parsers now additionally return the result of a parsed line. This enables live usage within a Gurobi Callback.
 
 ## 3.2.0 - 2024-12-19
 ### Fixed